Added comments and cleaned up code.

+++ src/table.cc
@@ -4,13 +4,12 @@
 
 #include "mylibrary/table.h"
 #include <cinder/app/App.h>
-#include <cinder/Log.h>
 #include <Box2D/Box2D.h>
-#include <cinder/app/App.h>
 #include <cinder/gl/gl.h>
 #include "mylibrary/box2d_utility.h"
 
 using namespace cinder;
+
 namespace myapp {
     Wall::Wall(BodyRef body)
             : Body(body, nullptr, vec2(0, 0))
@@ -24,7 +23,6 @@ namespace myapp {
     void Wall::draw() {
         float pointsPerMeter = Box2DUtility::getPointsPerMeter();
         auto pos = pointsPerMeter * body_->GetPosition();
-
         Rectf wallRect( pos.x - width_, 0.0f, pos.x + width_, app::getWindowHeight() );
         gl::drawSolidRect( wallRect );
 
@@ -33,7 +31,6 @@ namespace myapp {
         : Body(body, texture_, pos)
         {
             body_->SetUserData(this);
-            //SetPockets();
     }
     void Table::draw() {
         cinder::Area area = texture_->getBounds();
@@ -42,12 +39,12 @@ namespace myapp {
         cinder::gl::draw(texture_, fit);
     }
     void Table::SetPockets() {
-
+        // hard-coded numbers of positions of pockets
         auto posX = (app::getWindowWidth() - texture_->getWidth()) / 2;
-        float right = app::getWindowWidth() - 100.0f;
-        float left = app::getWindowWidth() - 710.0f;
-        float top = app::getWindowHeight() - 230.0f;
-        float bottom = app::getWindowHeight() - 525.0f;
+        float right = (float)app::getWindowWidth() - 100.0f;
+        float left = (float)app::getWindowWidth() - 710.0f;
+        float top = (float)app::getWindowHeight() - 230.0f;
+        float bottom = (float)app::getWindowHeight() - 525.0f;
 
         pockets_.push_back(vec2(left, bottom)); // bottom left
         pockets_.push_back(vec2(left, top)); // top left
@@ -63,15 +60,12 @@ namespace myapp {
             float x = Box2DUtility::metersToPoints(b->getPos().x );
             float y = Box2DUtility::metersToPoints(b->getPos().y );
             double d = std::hypot(std::abs(x - p.x), std::abs(y - p.y));
-            //float d = Box2DUtility::pointsToMeters(dist);
-            //float thresh = Box2DUtility::pointsToMeters(b->getRadius());
             float thresh = b->getRadius();
-           // float thresh = 50;
 
+            // check is ball's position is its radius distance away from pocket
             if (d <= thresh) // hit the pocket
                 return true;
         }
-
         return false;
     }
     void Table::handleCollision(Ball *ball, const ci::vec2 &contactPoint ) {
